+++ server/exchangers_test.go
@@ -11,6 +11,9 @@ func TestResourcesFulfillInterfaces(t *testing.T) {
 	assert.Implements(t, (*restful.Getable)(nil), newGDMResource(ComponentLocator{}))
 	assert.Implements(t, (*restful.Putable)(nil), newGDMResource(ComponentLocator{}))
 
+	assert.Implements(t, (*restful.Getable)(nil), newStateDeploymentResource(ComponentLocator{}))
+	assert.Implements(t, (*restful.Putable)(nil), newStateDeploymentResource(ComponentLocator{}))
+
 	assert.Implements(t, (*restful.Getable)(nil), newStateDefResource(ComponentLocator{}))
 
 	assert.Implements(t, (*restful.Getable)(nil), newManifestResource(ComponentLocator{}))
@@ -25,4 +28,10 @@ func TestResourcesFulfillInterfaces(t *testing.T) {
 	assert.Implements(t, (*restful.Getable)(nil), newStatusResource(ComponentLocator{}))
 
 	assert.Implements(t, (*restful.Getable)(nil), newHealthResource(ComponentLocator{}))
+
+	assert.Implements(t, (*restful.Getable)(nil), newAllDeployQueuesResource(ComponentLocator{}))
+
+	assert.Implements(t, (*restful.Getable)(nil), newDeployQueueResource(ComponentLocator{}))
+
+	assert.Implements(t, (*restful.Getable)(nil), newR11nResource(ComponentLocator{}))
 }