+++ server/handle_alldeployqueues.go
@@ -25,7 +25,7 @@ func newAllDeployQueuesResource(ctx ComponentLocator) *AllDeployQueuesResource {
 }
 
 // Get returns a configured GETAllDeployQueuesHandler.
-func (r *AllDeployQueuesResource) Get(_ http.ResponseWriter, _ *http.Request, _ httprouter.Params) restful.Exchanger {
+func (r *AllDeployQueuesResource) Get(_ *restful.RouteMap, _ http.ResponseWriter, _ *http.Request, _ httprouter.Params) restful.Exchanger {
 	return &GETAllDeployQueuesHandler{
 		QueueSet: r.context.QueueSet,
 	}